requirements.txt
dropbox
pandas
tabulate
emoji
psycopg2-binary
pymysql
requests
discord.py[voice]
prettytable
sqlalchemy
flask